Fixes for windows warnings savannah 15808
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@40890 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 467a6e59a1fe36f4f58f42102897e1e738d04795

+++ Rec/Calo/CaloMoniDst/src/CaloElectronNtp.cpp
@@ -1,4 +1,4 @@
-// $Id: CaloElectronNtp.cpp,v 1.1 2010-05-20 09:55:38 odescham Exp $
+// $Id: CaloElectronNtp.cpp,v 1.2 2010-06-03 22:21:46 rlambert Exp $
 // Include files 
 
 // from Gaudi
@@ -83,7 +83,7 @@ StatusCode CaloElectronNtp::execute() {
 
   // GET ODIN INFO
   int run = 0;
-  long evt = 0;
+  unsigned long long int evt = 0;
   int tty = 0;
   m_odin->getTime();
   if( exist<LHCb::ODIN>(LHCb::ODINLocation::Default) ){
